fix workspace ownership issue

+++ routes/workspace/index.js
@@ -6,6 +6,7 @@ const {
     deleteWorkspace,
     getOneWorkspace,
 } = require("../../controllers/workspace");
+const { getLoggedinUID, hetHasOwnership } = require("../../utils");
 
 const router = Router();
 
@@ -30,8 +31,8 @@ router.get("/", async (req, res) => {
 
 router.get("/:id", async (req, res) => {
     try {
+        const uid = getLoggedinUID(req.auth);
         const { id } = req.params;
-        const { uid } = req.query;
 
         if (id && uid) {
             const data = await getOneWorkspace(id, uid);
@@ -46,13 +47,19 @@ router.get("/:id", async (req, res) => {
 
 router.post("/new", async (req, res) => {
     try {
-        const { creatorId, clientId, clientSecret } = req.body;
+        const isOwner = hetHasOwnership(req.auth, req.body.creatorId);
 
-        if (creatorId && clientId && clientSecret) {
-            const { status, workspace } = await addNewWorkspace(req.body);
-            res.status(status).json(workspace);
+        if (isOwner) {
+            const { creatorId, clientId, clientSecret } = req.body;
+
+            if (creatorId && clientId && clientSecret) {
+                const { status, workspace } = await addNewWorkspace(req.body);
+                res.status(status).json(workspace);
+            } else {
+                res.status(400).send();
+            }
         } else {
-            res.status(400).send();
+            res.status(401).send();
         }
     } catch (error) {
         res.status(500).send();
@@ -61,14 +68,20 @@ router.post("/new", async (req, res) => {
 
 router.put("/edit/:id", async (req, res) => {
     try {
-        const { id } = req.params;
-        const { clientId, clientSecret } = req.body;
+        const isOwner = hetHasOwnership(req.auth, req.body.creatorId);
 
-        if (id && clientId && clientSecret) {
-            const { status, workspace } = await editWorkspace(req.body, id);
-            res.status(status).json(workspace);
+        if (isOwner) {
+            const { id } = req.params;
+            const { clientId, clientSecret } = req.body;
+
+            if (id && clientId && clientSecret) {
+                const { status, workspace } = await editWorkspace(req.body, id);
+                res.status(status).json(workspace);
+            } else {
+                res.status(400).send();
+            }
         } else {
-            res.status(400).send();
+            res.status(401).send();
         }
     } catch (error) {
         res.status(500).send();
@@ -77,12 +90,19 @@ router.put("/edit/:id", async (req, res) => {
 
 router.delete("/delete/:id", async (req, res) => {
     try {
-        const { id } = req.params;
-        if (id) {
-            await deleteWorkspace(id);
-            res.status(200).send();
+        const isOwner = hetHasOwnership(req.auth, req.body.creatorId);
+
+        if (isOwner) {
+            const { id } = req.params;
+
+            if (id) {
+                await deleteWorkspace(id);
+                res.status(200).send();
+            } else {
+                res.status(400).send();
+            }
         } else {
-            res.status(400).send();
+            res.status(401).send();
         }
     } catch (error) {
         res.status(500).send();
